% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucnrle.R
\name{engine_version}
\alias{engine_version}
\title{Version of the underlying calculation engine}
\usage{
engine_version()
}
\value{
A version string.
}
\description{
Version of the underlying calculation engine
}
