% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rle_version}
\alias{rle_version}
\title{Version of the underlying iucn-rle-core calculation engine.}
\usage{
rle_version()
}
\description{
Version of the underlying iucn-rle-core calculation engine.
}
